requirements.txt
beautifulsoup4==4.9.1
certifi==2020.6.20
chardet==3.0.4
contextvars==2.4
fire==0.3.1
googletrans==3.0.0
h11==0.9.0
h2==3.2.0
hpack==3.0.0
hstspreload==2020.9.15
httpcore==0.9.1
httpx==0.13.3
hyperframe==5.2.0
idna==2.10
immutables==0.14
lxml==4.5.2
pdf2docx==0.4.2
PyMuPDF==1.17.7
python-docx==0.8.10
requests==2.24.0
rfc3986==1.4.0
six==1.15.0
sniffio==1.1.0
soupsieve==2.0.1
termcolor==1.1.0
tqdm==4.49.0
urllib3==1.25.10
